requirement.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
dateparser==1.1.3
discord.py==2.0.1
frozenlist==1.3.1
geographiclib==1.52
geopy==2.2.0
GoogleNews==1.6.5
h3==3.7.4
idna==3.4
lavalink==4.0.4
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
timezonefinder==6.1.6
tweepy==4.12.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wavelink==1.3.3
yarl==1.8.1
